Abstract API request to new class
* Create Singleton NetworkHandler class
* Make NetworkHandler responsible of tracking request
