+++ lib/osproject.rb
@@ -32,7 +32,7 @@ class OSProject
   end
 
   def self.version
-    '1.0'
+    '0.0.0'
   end
 
   def self.os
